fix(core): guard Win-only paths and 64-bit cast pinch-points across GameEngine
(cherry picked from commit d766abb0dad7997dd50b74c9fc9b8fc038deba68)

+++ Core/GameEngine/Source/Common/System/LocalFile.cpp
@@ -370,7 +370,7 @@ Int LocalFile::readChar()
 	Int ret = read( &character, sizeof(character) );
 
 	if (ret == sizeof(character))
-		return (Int)character;
+		return (Int)(intptr_t)character;
 
 	return EOF;
 }
@@ -386,7 +386,7 @@ Int LocalFile::readWideChar()
 	Int ret = read( &character, sizeof(character) );
 
 	if (ret == sizeof(character))
-		return (Int)character;
+		return (Int)(intptr_t)character;
 
 	return WEOF;
 }
@@ -450,7 +450,7 @@ Int LocalFile::writeFormat( const WideChar* format, ... )
 Int LocalFile::writeChar( const Char* character )
 {
 	if ( write( character, sizeof(Char) ) == sizeof(Char) ) {
-		return (Int)character;
+		return (Int)(intptr_t)character;
 	}
 
 	return EOF;
@@ -463,7 +463,7 @@ Int LocalFile::writeChar( const Char* character )
 Int LocalFile::writeChar( const WideChar* character )
 {
 	if ( write( character, sizeof(WideChar) ) == sizeof(WideChar) ) {
-		return (Int)character;
+		return (Int)(intptr_t)character;
 	}
 
 	return WEOF;
